refactor: reorganized command handlers
