+++ client/src/lib.rs
@@ -1,5 +1,3 @@
-use common::configs::model_config::ModelIndex;
-use mesh_color::MeshColor;
 use glm::vec3;
 use other_players::OtherPlayer;
 
@@ -21,9 +19,9 @@ mod model;
 mod camera;
 mod instance;
 mod lights;
+mod other_players;
 mod particles;
 mod player;
-mod other_players;
 mod resources;
 mod scene;
 mod screen;
@@ -41,9 +39,7 @@ use crate::animation::AnimatedModel;
 use crate::inputs::Input;
 use crate::model::{Model, StaticModel};
 
-use common::configs::model_config::ConfigModels;
-use common::configs::scene_config::ConfigSceneGraph;
-use common::configs as configs;
+use common::configs;
 use common::core::command::Command;
 use common::core::events;
 use common::core::states::{GameState, ParticleQueue};
@@ -253,22 +249,20 @@ impl State {
                 label: Some("texture_bind_group_layout"),
             });
 
-            let color_bind_group_layout =
-                device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-                    entries: &[
-                        wgpu::BindGroupLayoutEntry {
-                            binding: 0,
-                            visibility: wgpu::ShaderStages::FRAGMENT,
-                            ty: wgpu::BindingType::Buffer {
-                                ty: wgpu::BufferBindingType::Uniform,
-                                has_dynamic_offset: false,
-                                min_binding_size: None,
-                            },
-                            count: None,
-                        }
-                    ],
-                    label: Some("color_bind_group_layout"),
-                });
+        let color_bind_group_layout =
+            device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+                entries: &[wgpu::BindGroupLayoutEntry {
+                    binding: 0,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Buffer {
+                        ty: wgpu::BufferBindingType::Uniform,
+                        has_dynamic_offset: false,
+                        min_binding_size: None,
+                    },
+                    count: None,
+                }],
+                label: Some("color_bind_group_layout"),
+            });
 
         let texture_bind_group_layout_2d =
             device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
@@ -491,9 +485,10 @@ impl State {
         let glyph_brush = GlyphBrushBuilder::using_font(inconsolata).build(&device, surface_format);
 
         let rng = rand::thread_rng();
-        let particle_tex = texture::Texture::from_images(&config_instance.texture.particles, &device, &queue)
-            .await
-            .unwrap();
+        let particle_tex =
+            texture::Texture::from_images(&config_instance.texture.particles, &device, &queue)
+                .await
+                .unwrap();
         let particle_renderer = particles::ParticleDrawer::new(
             &device,
             &config,
@@ -511,7 +506,7 @@ impl State {
                 // TODO: skipping animated models for now for lobby scene
                 continue;
             } else {
-                let model: Box<dyn Model> =  Box::new(
+                let model: Box<dyn Model> = Box::new(
                     StaticModel::load(&model_config.path, model_loading_resources)
                         .await
                         .unwrap(),
@@ -519,9 +514,9 @@ impl State {
                 models.insert(model_config.name, model);
             }
         }
-        
+
         let lobby_scene_config = config_instance.lobby_scene.clone();
-        
+
         let mut lobby_scene = scene::Scene::from_config(&lobby_scene_config);
         lobby_scene.objects = models;
         lobby_scene.draw_scene_dfs();
@@ -532,7 +527,6 @@ impl State {
         scene_map.insert(String::from("scene:game"), scene);
         scene_map.insert(String::from("scene:lobby"), lobby_scene);
 
-        
         // end debug code that needs to be replaced
 
         let mut texture_map: HashMap<String, wgpu::BindGroup> = HashMap::new();
@@ -580,36 +574,33 @@ impl State {
             sender,
             game_state,
         );
-        
-        let other_players: Vec<OtherPlayer> = (1..5)
-            .map(|ind| 
-                OtherPlayer{
-                    id: ind,
-                    visible: false,
-                    location: glm::vec4(0.0, 0.0, 0.0, 0.0),
-                    score: 0.0,
-                }
-            ).collect();
+
+        let _other_players: Vec<OtherPlayer> = (1..5)
+            .map(|ind| OtherPlayer {
+                id: ind,
+                visible: false,
+                location: glm::vec4(0.0, 0.0, 0.0, 0.0),
+                score: 0.0,
+            })
+            .collect();
+
+        let _other_players: Vec<OtherPlayer> = (1..5)
+            .map(|ind| OtherPlayer {
+                id: ind,
+                visible: false,
+                location: glm::vec4(0.0, 0.0, 0.0, 0.0),
+                score: 0.0,
+            })
+            .collect();
 
         let other_players: Vec<OtherPlayer> = (1..5)
-            .map(|ind| 
-                OtherPlayer{
-                    id: ind,
-                    visible: false,
-                    location: glm::vec4(0.0, 0.0, 0.0, 0.0),
-                    score: 0.0,
-                }
-            ).collect();
-
-        let other_players: Vec<OtherPlayer> = (1..5)
-            .map(|ind| 
-                OtherPlayer{
-                    id: ind,
-                    visible: false,
-                    location: glm::vec4(0.0, 0.0, 0.0, 0.0),
-                    score: 0.0,
-                }
-            ).collect();
+            .map(|ind| OtherPlayer {
+                id: ind,
+                visible: false,
+                location: glm::vec4(0.0, 0.0, 0.0, 0.0),
+                score: 0.0,
+            })
+            .collect();
 
         Self {
             window,
@@ -727,93 +718,136 @@ impl State {
                     dt,
                     self.client_id,
                 );
-            
+
             other_players::load_game_state(&mut self.other_players, game_state.lock().unwrap());
-            
+
             // update player scores
             // PLACEHOLDER FOR NOW
             {
-                let screen_id = self.display.groups
-                .get(&self.display.game_display)
-                .unwrap()
-                .screen
-                .as_ref()
-                .unwrap();
+                let screen_id = self
+                    .display
+                    .groups
+                    .get(&self.display.game_display)
+                    .unwrap()
+                    .screen
+                    .as_ref()
+                    .unwrap();
 
                 let screen = self.display.screen_map.get_mut(screen_id).unwrap();
-                for i in 1..5{
-                    let ind = screen.icon_id_map.get(&format!("icon:score_p{}",i)).unwrap().clone();
-                    let score : f32 = self.other_players[i as usize - 1].score;
-                    let mut location = screen.icons[ind].location.clone();
-                    location.horz_disp = (0.0, parameters::SCORE_LOWER_X + score * (parameters::SCORE_UPPER_X - parameters::SCORE_LOWER_X));
+                for i in 1..5 {
+                    let ind = *screen
+                        .icon_id_map
+                        .get(&format!("icon:score_p{}", i))
+                        .unwrap();
+                    let score: f32 = self.other_players[i as usize - 1].score;
+                    let mut location = screen.icons[ind].location;
+                    location.horz_disp = (
+                        0.0,
+                        parameters::SCORE_LOWER_X
+                            + score * (parameters::SCORE_UPPER_X - parameters::SCORE_LOWER_X),
+                    );
                     screen.icons[ind].relocate(
                         location,
                         self.config.width,
                         self.config.height,
-                        &self.queue
+                        &self.queue,
                     );
                 }
             }
 
             // update player number of charges
             {
-                let screen_id = self.display.groups
-                .get(&self.display.game_display)
-                .unwrap()
-                .screen
-                .as_ref()
-                .unwrap();
+                let screen_id = self
+                    .display
+                    .groups
+                    .get(&self.display.game_display)
+                    .unwrap()
+                    .screen
+                    .as_ref()
+                    .unwrap();
 
                 let screen = self.display.screen_map.get_mut(screen_id).unwrap();
-                let ind = screen.icon_id_map.get("icon:charge").unwrap().clone();
+                let ind = *screen.icon_id_map.get("icon:charge").unwrap();
                 screen.icons[ind].inst_range = 0..self.player.wind_charge;
             }
 
             // update cooldowns
             // hard coded for now... TODO: separate function
             // is it necessary? would need to pass around lots of references
-            // might be better to create dedicated function in screen/mod.rs
+            // might be better to create dedicated function in screen/command_handlers
             {
-                let screen_id = self.display.groups
-                .get(&self.display.game_display)
-                .unwrap()
-                .screen
-                .as_ref()
-                .unwrap();
+                let screen_id = self
+                    .display
+                    .groups
+                    .get(&self.display.game_display)
+                    .unwrap()
+                    .screen
+                    .as_ref()
+                    .unwrap();
 
                 let screen = self.display.screen_map.get_mut(screen_id).unwrap();
-                let ind = screen.icon_id_map.get("icon:atk_forward_overlay").unwrap().clone();
+                let ind = *screen.icon_id_map.get("icon:atk_forward_overlay").unwrap();
                 let mut tint;
 
                 if self.player.on_cooldown.contains_key(&Command::Attack) {
-                    let cd_left = self.player.on_cooldown.get(&Command::Attack).unwrap() / common::configs::parameters::ATTACK_COOLDOWN;
+                    let cd_left = self.player.on_cooldown.get(&Command::Attack).unwrap()
+                        / common::configs::parameters::ATTACK_COOLDOWN;
                     // use smmoothstep?
-                    screen.icons[ind].tint = glm::vec4(1.0, 1.0, 1.0, 3.0 * cd_left.powf(2.0) - 2.0 * cd_left.powf(3.0));
-                    tint = glm::vec4(1.0, 1.0, 1.0, 3.0 * cd_left.powf(2.0) - 2.0 * cd_left.powf(3.0));
+                    screen.icons[ind].tint = glm::vec4(
+                        1.0,
+                        1.0,
+                        1.0,
+                        3.0 * cd_left.powf(2.0) - 2.0 * cd_left.powf(3.0),
+                    );
+                    tint = glm::vec4(
+                        1.0,
+                        1.0,
+                        1.0,
+                        3.0 * cd_left.powf(2.0) - 2.0 * cd_left.powf(3.0),
+                    );
                 } else {
                     screen.icons[ind].tint = glm::vec4(1.0, 1.0, 1.0, 0.0);
                     tint = glm::vec4(1.0, 1.0, 1.0, 0.0);
                 }
-                for v in &mut screen.icons[ind].vertices{
+                for v in &mut screen.icons[ind].vertices {
                     v.color = tint.into();
                 }
-                self.queue.write_buffer(&screen.icons[ind].vbuf, 0, bytemuck::cast_slice(&screen.icons[ind].vertices));
+                self.queue.write_buffer(
+                    &screen.icons[ind].vbuf,
+                    0,
+                    bytemuck::cast_slice(&screen.icons[ind].vertices),
+                );
 
-                let ind = screen.icon_id_map.get("icon:atk_wave_overlay").unwrap().clone();
+                let ind = *screen.icon_id_map.get("icon:atk_wave_overlay").unwrap();
 
                 if self.player.on_cooldown.contains_key(&Command::AreaAttack) {
-                    let cd_left = self.player.on_cooldown.get(&Command::AreaAttack).unwrap() / common::configs::parameters::AREA_ATTACK_COOLDOWN;
+                    let cd_left = self.player.on_cooldown.get(&Command::AreaAttack).unwrap()
+                        / common::configs::parameters::AREA_ATTACK_COOLDOWN;
                     // use smmoothstep?
-                    screen.icons[ind].tint = glm::vec4(1.0, 1.0, 1.0, 3.0 * cd_left.powf(2.0) - 2.0 * cd_left.powf(3.0));
-                    tint = glm::vec4(1.0, 1.0, 1.0, 3.0 * cd_left.powf(2.0) - 2.0 * cd_left.powf(3.0));
+                    screen.icons[ind].tint = glm::vec4(
+                        1.0,
+                        1.0,
+                        1.0,
+                        3.0 * cd_left.powf(2.0) - 2.0 * cd_left.powf(3.0),
+                    );
+                    tint = glm::vec4(
+                        1.0,
+                        1.0,
+                        1.0,
+                        3.0 * cd_left.powf(2.0) - 2.0 * cd_left.powf(3.0),
+                    );
                 } else {
                     screen.icons[ind].tint = glm::vec4(1.0, 1.0, 1.0, 0.0);
                     tint = glm::vec4(1.0, 1.0, 1.0, 0.0);
                 }
-                for v in &mut screen.icons[ind].vertices{
+                for v in &mut screen.icons[ind].vertices {
                     v.color = tint.into();
                 }
-                self.queue.write_buffer(&screen.icons[ind].vbuf, 0, bytemuck::cast_slice(&screen.icons[ind].vertices));
+                self.queue.write_buffer(
+                    &screen.icons[ind].vbuf,
+                    0,
+                    bytemuck::cast_slice(&screen.icons[ind].vertices),
+                );
             }
 
             self.display
@@ -830,10 +864,10 @@ impl State {
                 .get_player_positions();
 
             // ASSUME: Ids should always be 1-4
-            for p in &mut self.other_players{
+            for p in &mut self.other_players {
                 p.visible = false;
             }
-            for (i, loc) in player_loc{
+            for (i, loc) in player_loc {
                 self.other_players[i as usize - 1].location = loc;
                 self.other_players[i as usize - 1].visible = true;
             }
@@ -905,7 +939,7 @@ impl State {
                     Text::new("Respawning in ")
                         .with_color([1.0, 1.0, 0.0, 1.0])
                         .with_scale(60.0),
-                    Text::new(&format!("{:.1}", spawn_cooldown).as_str())
+                    Text::new(format!("{:.1}", spawn_cooldown).as_str())
                         .with_color([1.0, 1.0, 1.0, 1.0])
                         .with_scale(60.0),
                     Text::new(" seconds")
@@ -931,11 +965,11 @@ impl State {
         self.glyph_brush.queue(Section {
             screen_position: (600.0, 60.0),
             bounds: (size.width as f32, size.height as f32),
-            text: vec![Text::new(
-                format!("PowerUp Held: {:?}\n", self.player.power_up).as_str(),
-            )
-            .with_color([0.0, 0.0, 0.0, 1.0])
-            .with_scale(40.0)],
+            text: vec![
+                Text::new(format!("PowerUp Held: {:?}\n", self.player.power_up).as_str())
+                    .with_color([0.0, 0.0, 0.0, 1.0])
+                    .with_scale(40.0),
+            ],
             ..Section::default()
         });
 
@@ -998,7 +1032,7 @@ impl State {
                         &mut self.rng,
                     );
                     self.display.particles.systems.push(atk);
-                },
+                }
                 events::ParticleType::AREA_ATTACK => {
                     // in this case, only position matters
                     let time = parameters::AREA_ATTACK_COOLDOWN / 2.0;
