+++ client/src/mesh_color.rs
@@ -7,15 +7,13 @@ pub struct MeshColor {
 }
 
 impl MeshColor {
-    pub fn new(rgb: [f32; 3]) -> Self{
-        MeshColor {
-            rgb_color: rgb,
-        }
+    pub fn new(rgb: [f32; 3]) -> Self {
+        MeshColor { rgb_color: rgb }
     }
-    
+
     pub fn default() -> Self {
         MeshColor {
-            rgb_color: [1.0,1.0,1.0],
+            rgb_color: [1.0, 1.0, 1.0],
         }
     }
 }
@@ -40,7 +38,7 @@ impl MeshColorInstance {
         });
 
         let color_bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            layout: &color_bind_group_layout,
+            layout: color_bind_group_layout,
             entries: &[wgpu::BindGroupEntry {
                 binding: 0,
                 resource: color_buffer.as_entire_binding(),
