+++ client/src/screen/display_helper.rs
@@ -1,3 +1,4 @@
+use crate::mesh_color::{MeshColor, MeshColorInstance};
 use crate::screen::location_helper::{get_coords, to_absolute};
 use crate::screen::objects;
 use crate::screen::objects::ScreenInstance;
@@ -5,7 +6,6 @@ use common::configs::display_config::{ConfigDisplay, ConfigScreen};
 use nalgebra_glm as glm;
 use std::collections::HashMap;
 use wgpu::util::DeviceExt;
-use crate::mesh_color::{MeshColorInstance, MeshColor};
 
 pub fn create_display_group(config: &ConfigDisplay) -> HashMap<String, objects::DisplayGroup> {
     let mut groups = HashMap::new();
@@ -30,9 +30,15 @@ pub fn create_screen_map(
     let mut screen_map = HashMap::new();
     for s in &config.screens {
         let background = create_background(s, device, color_bind_group_layout);
-        let mut icon_id_map : HashMap<String, usize> = HashMap::new();
+        let mut icon_id_map: HashMap<String, usize> = HashMap::new();
         let icons = create_icon(s, device, &mut icon_id_map, screen_width, screen_height);
-        let buttons = create_button(s, device, screen_width, screen_height, color_bind_group_layout);
+        let buttons = create_button(
+            s,
+            device,
+            screen_width,
+            screen_height,
+            color_bind_group_layout,
+        );
 
         let screen = objects::Screen {
             id: s.id.clone(),
@@ -40,14 +46,22 @@ pub fn create_screen_map(
             icons,
             icon_id_map,
             buttons,
-            default_color:  MeshColorInstance::new(device, color_bind_group_layout, MeshColor::default()),
+            default_color: MeshColorInstance::new(
+                device,
+                color_bind_group_layout,
+                MeshColor::default(),
+            ),
         };
         screen_map.insert(s.id.clone(), screen);
     }
     screen_map
 }
 
-fn create_background(s: &ConfigScreen, device: &wgpu::Device, color_bind_group_layout: &wgpu::BindGroupLayout) -> Option<objects::ScreenBackground> {
+fn create_background(
+    s: &ConfigScreen,
+    device: &wgpu::Device,
+    color_bind_group_layout: &wgpu::BindGroupLayout,
+) -> Option<objects::ScreenBackground> {
     s.background.as_ref().map(|bg| {
         let vertices = objects::TITLE_VERT;
         let vbuf = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
@@ -60,7 +74,9 @@ fn create_background(s: &ConfigScreen, device: &wgpu::Device, color_bind_group_l
             aspect: bg.aspect,
             vbuf,
             texture: bg.tex.clone(),
-            color: match bg.color {None => None, Some(c) => Some(MeshColorInstance::new(device, color_bind_group_layout, MeshColor::new(c)))},
+            color: bg.color.map(|c| {
+                MeshColorInstance::new(device, color_bind_group_layout, MeshColor::new(c))
+            }),
         }
     })
 }
@@ -87,8 +103,8 @@ fn create_icon(
                 screen_height,
                 &mut vertices,
             );
-            for v in &mut vertices{
-                v.color = i.tint.clone();
+            for v in &mut vertices {
+                v.color = i.tint;
             }
             let vbuf = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
                 label: Some(" Some icon (implement ids for more useful messages!) Vertex Buffer"),
@@ -163,7 +179,7 @@ fn create_button(
             });
 
             objects::Button {
-                id: b.id.clone(), 
+                id: b.id.clone(),
                 location: b.location,
                 aspect: b.aspect,
                 height: b.height,
@@ -173,7 +189,9 @@ fn create_button(
                 hover_tint: glm::make_vec4(&b.hover_tint),
                 default_texture: b.default_tex.clone(),
                 hover_texture: b.hover_tex.clone(),
-                color: match b.color{None => None, Some(c) => Some(MeshColorInstance::new(device, color_bind_group_layout, MeshColor::new(c)))},
+                color: b.color.map(|c| {
+                    MeshColorInstance::new(device, color_bind_group_layout, MeshColor::new(c))
+                }),
                 on_click: b.on_click.clone(),
             }
         })
