+++ server/src/executor/command_handlers/weather.rs
@@ -1,0 +1,1 @@
+
