+++ server/src/executor/mod.rs
@@ -5,12 +5,6 @@ use common::core::command::{Command, MoveDirection, ServerSync};
 use common::core::events::GameEvent;
 use common::core::states::GameState;
 
-use crate::executor::command_handlers::{
-    AttackCommandHandler, CastPowerUpCommandHandler, CommandHandler, DashCommandHandler,
-    DieCommandHandler, FlashCommandHandler, JumpCommandHandler, MoveCommandHandler,
-    RefillCommandHandler, SpawnCommandHandler, StartupCommandHandler,
-    UpdateCameraFacingCommandHandler, AreaAttackCommandHandler,
-};
 use crate::game_loop::ClientCommand;
 use crate::simulation::physics_state::PhysicsState;
 use crate::Recipients;
@@ -18,8 +12,10 @@ use crate::Recipients;
 use common::core::states::GameLifeCycleState::{Running, Waiting};
 use itertools::Itertools;
 use log::{debug, error, info, warn};
-use std::cell::{RefCell, RefMut};
+use std::cell::RefCell;
 use std::time::Duration;
+
+use command_handlers::prelude::*;
 
 pub mod command_handlers;
 
@@ -162,7 +158,9 @@ impl Executor {
                 ),
                 Command::Jump => Box::new(JumpCommandHandler::new(client_command.client_id)),
                 Command::Attack => Box::new(AttackCommandHandler::new(client_command.client_id)),
-                Command::AreaAttack => Box::new(AreaAttackCommandHandler::new(client_command.client_id)),
+                Command::AreaAttack => {
+                    Box::new(AreaAttackCommandHandler::new(client_command.client_id))
+                }
                 Command::Refill => Box::new(RefillCommandHandler::new(client_command.client_id)),
                 Command::CastPowerUp => {
                     Box::new(CastPowerUpCommandHandler::new(client_command.client_id))
@@ -246,17 +244,5 @@ impl Executor {
     /// get a clone of the game state
     pub fn game_state(&self) -> GameState {
         self.game_state.lock().unwrap().clone()
-    }
-}
-
-type GameEventWithRecipients = (GameEvent, Recipients);
-
-pub trait GameEventCollector {
-    fn add(&mut self, event: GameEvent, recipients: Recipients);
-}
-
-impl GameEventCollector for RefMut<'_, Vec<GameEventWithRecipients>> {
-    fn add(&mut self, event: GameEvent, recipients: Recipients) {
-        self.push((event, recipients));
     }
 }
