+++ server/src/main.rs
@@ -10,10 +10,11 @@ use std::sync::atomic::AtomicBool;
 use std::sync::{mpsc, Arc, Mutex};
 use std::{net::TcpListener, thread};
 
+#[allow(unused_imports)]
 use common::communication::commons::CSE125_SERVER_ADDR;
+#[allow(unused_imports)]
 use common::communication::commons::DEFAULT_SERVER_ADDR;
 
-use common::configs::*;
 use server::executor::Executor;
 use threadpool::ThreadPool;
